pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["dhiulia <antunesdhiulia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
drf-spectacular = "^0.22.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
